✨ feat(TaskCell.swift) | ✨ feat(TaskListVC.swift)
The commit removes unnecessary iCloud files from the RevealServer directory. The Helpers.swift file now has a new function called formatDateWithoutYear that formats a date without the year. The TaskCell.swift file now has a new deadlineStackView, calendarImageView, and deadlineLabel to display the task deadline. The statusButton is added to toggle the task status. The Task objects in TaskListVC have been updated to improve readability. The changes include updating the deadline of one task and removing the description of another task. A header view has been added to TasksList to display the number of completed tasks and a hide/show button. The header view is displayed at the top of the table view and can be used to toggle the visibility of completed tasks. The TasksList table view has also been updated to use the .insetGrouped style instead of the .plain style to improve the appearance of the table view.