🎨 style(Fonts.swift) | 🐛 fix(DataManager.swift) | 🐛 fix(FileCache.swift) | 🐛 fix(DeleteButton.swift) | 🎨 style(TaskDetailsVC.swift) | ✨ feat(TaskDetailsVC.swift) | 🐛 fix(TaskCell.swift) | 🔥 chore(TaskListVC.swift) | ✨ feat(TaskListVC.swift)
The changes in Fonts.swift improve the visual consistency of the app by decreasing the font size of the largeTitle. The changes in DataManager.swift and FileCache.swift fix a bug where the tasks dictionary was not being properly updated and was causing issues with the app's functionality. The changes in DeleteButton.swift enable the delete button by default to improve the user experience. The commented out code and unused variable were removed to improve code readability. The refreshView function was added to TaskDetailsVC to allow the view to be updated with a selected task. The prepareForReuse function was added to TaskCell to fix an issue where strikethrough text was not being removed when the cell was reused. The commented out code and unused properties have been removed to improve code readability. The fileCache property has been added to support loading and saving tasks to a JSON file. The getTasks() method has been added to retrieve tasks from the fileCache. The saveTask() and deleteTask() methods have been added to save and delete tasks from the fileCache. This allows the application to persist tasks between launches.