📦 chore(URLSession.swift)
The URLSession.swift file was added to the Common group in the project structure. The URLSession.swift file contains an extension of URLSession with a new method dataTask(for request: URLRequest) that supports asynchronous execution using async/await. This method allows for performing data tasks with a given URLRequest and returns a tuple of Data and URLResponse. The method also handles task cancellation.