🔧 fix(TaskDetails.swift) | 🔧 fix(Helpers.swift)
The date state variable is added to TaskDetails view to enable the selection of a date. The default value is set to Date() to show the current date as the initial value. In the formatDate function, the locale is set to ru_RU to format dates in the Russian language. This ensures that the dates are displayed correctly in the desired language.